updated registration link and CENTRA 3 mtg title
Nadya - I set up the PRAGMA / CENTRA 2018 Meetings registration for for the events in May and admin the globalcentra.org/centra3 content. Proposed change is for consistency when CENTRA 3 Meeting is mentioned - with 'Smart Cyberinfrastructure for Transnational Science' and removed 'All-Hands'. Shoot me an email if any question/concern or a quick Skype call is more effective. Thank you.